+++ core-network/src/main/java/com/jik/core/network/source/MovieRemoteDataSource.kt
@@ -5,7 +5,7 @@ import com.jik.core.model.MovieInfo
 
 interface MovieRemoteDataSource {
 
-    suspend fun getPopularMovies(page:Int): Result<List<Movie>>
+    suspend fun getPopularMovies(page: Int): Result<List<Movie>>
 
-    suspend fun getMovieInfo(id: Int): Result<MovieInfo>
+    suspend fun getMovieInfo(id: Long): Result<MovieInfo>
 }
